Finish diagnose and fix
